requirements.txt
streamlit>=1.54.0
openai>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
